Updating new search component for front end changes
